+++ tests/object.rs
@@ -295,7 +295,7 @@ async fn test_optional_output_with_try() {
 
     #[Object]
     impl Query {
-        async fn obj(&self, ctx: &Context<'_>) -> Option<u32> {
+        async fn obj(&self, ctx: &Context<'_>) -> Option<u16> {
             let x = ctx.data::<B>().unwrap();
 
             if x.some? { Some(300) } else { None }
